Refatorando CpfCnpjUtils
