+++ grails-app/services/gocharges/PayerService.groovy
@@ -2,7 +2,7 @@ package gocharges
 
 import gocharges.payer.PayerRepository
 import gocharges.payer.adapter.PayerAdapter
-import gocharges.validator.CpfCnpjValidator
+import shared.CpfCnpjUtils
 import grails.gorm.transactions.Transactional
 import gocharges.exception.BusinessException
 
@@ -41,7 +41,7 @@ class PayerService {
 
     private void validateSave(PayerAdapter adapter)  {
         validateNotNull(adapter)
-        CpfCnpjValidator.validate(adapter.cpfCnpj)
+        CpfCnpjUtils.validate(adapter.cpfCnpj)
 
         Payer payer = PayerRepository.query([email: adapter.email, includeDeleted: true]).get()
         if(payer)  throw new BusinessException("Email já cadastrado.")
