+++ grails-app/utils/shared/Utils.groovy
@@ -1,0 +1,8 @@
+package shared
+
+public class Utils {
+    public static String removeNonNumeric(String text) {
+        if (text == null) return null
+        return text?.replaceAll("\\D+","")
+    }
+}
