+++ src/test/groovy/gocharges/DashboardServiceSpec.groovy
@@ -1,0 +1,18 @@
+package gocharges
+
+import grails.testing.services.ServiceUnitTest
+import spock.lang.Specification
+
+class DashboardServiceSpec extends Specification implements ServiceUnitTest<DashboardService>{
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
